+++ backend/pkg/server/server.go
@@ -81,6 +81,7 @@ func Serve(addr string) {
 	api := e.Group("/api")
 	api.POST("/user/create", userHandler.HandleUserCreate())
 	api.POST("/user/login", userHandler.HandleUserLogin())
+	
 
 	apiAuth := api.Group("/auth")
 	apiAuth.Use(mw.Authenticate)
@@ -92,6 +93,7 @@ func Serve(addr string) {
 	// img
 	apiAuth.POST("/trip/save", tripHandler.HandleTripSave())
 	apiAuth.GET("/trip/get", tripHandler.HandleTripGet())
+	apiAuth.GET("/trip/get_img", tripHandler.HandleImgGet())
 	// html
 	apiAuth.Static("/html", "static")
